Case 1/requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.22.0
opencv-python>=4.7.0
pillow>=9.5.0
numpy>=1.24.0